### 2019-09-16 #1735 @whedon
[PRE REVIEW]: Shop-Solver: A framework for solving and benchmarking shop scheduling problems
**Submitting author:** @KingZee (<a href="http://orcid.org/0000-0002-6478-3628">Abdelhakim Zakkari</a>)
**Repository:** <a href="https://github.com/KingZee/shop-solver" target ="_blank">https://github.com/KingZee/shop-solver</a>
**Version:** v1.2.0
