### 2019-09-22 #1752 @nnadeau
@gkthiruvathukal @KingZee my initial impressions:

## Authorship

The second and third authors have made [no contributions to the repository](https://github.com/KingZee/shop-solver/graphs/contributors):

- Charif Mabrouki
- Alami Semma

## Significance

- Makespan optimization is a well-studied research topic.
- The presented software claims to be "feature" complete, but is not much more than a GUI wrapper around basic optimization.
- The paper is a survey of process optimization and does not state the significance of this repository, beyond being a "framework for other researchers to build and improve upon".
- Unfortunately, as a framework it is lacking:
  - There is no documentation of the API.
  - A significant number of functions and code are missing javadoc and comments.
  - It is not clear how the code architecture is structured; there is no clear model-view-controller.
  - Blocks of commented out code shows poor code health.
  - The core solver function `solveMakespan()` is a large if-else block, limiting extensibility.
  - The `QuickSolver` algorithm claims to be "heuristic", but appears to simply iterates through schedule permutations.
- The authors claim a restricted scope to "problems with fixed, known-beforehand processing times, that will execute on a single available machine, with no setup-times, due dates or release dates", thus defining a simplified Gantt chart and limiting the scope of future research extensibility.

## Issues

- https://github.com/KingZee/shop-solver/pull/2

## Initial Impressions

- While process optimization is an important research topic, this repository appears to be more of a GUI wrapper than an extensible framework for research.
- The lack of documentation impedes researchers from using the code.
- There is nothing novel that researchers would not code themselves when developing solutions.
- Significant work is required prior to being ready for review.
