### 2019-09-23 #1752 @KingZee
@gkthiruvathukal @nnadeau, thanks a lot for your efforts.

I have to say I am a bit confused, but really willing to discuss and improve on anything that needs work here. What do you mean by a "GUI wrapper"? A wrapper to what? My own code for solving schedules?

What I built here is a documented structure that allows any future development to be extended from the same abstract class. You should know every piece of code in my project is mine, and the GUI isnt a wrapper for a single solver, it's an extension I myself chose to add that is agnostic to whatever solver type is used, making the software extremely easy to fork and improve upon. The [Solver](https://github.com/KingZee/shop-solver/blob/master/src/main/java/scheduler/Solver.java)
 (and [Benchmark]( https://github.com/KingZee/shop-solver/blob/master/src/main/java/scheduler/Benchmark.java)) classes are made to handle solver work across multiple threads.

The [solvers](https://github.com/KingZee/shop-solver/tree/master/src/main/java/scheduler/solvers) package has 2 solver algorithms that I coded and optimised myself for all 3 types. Like I mentioned before and in the readme, I plan to add many more solver types based on the same class structure explained above.

So, I really dont understand @nnadeau's argument here, the software doesn't just render a GUI, it actually solves any given matrix. The GUI is even more extra effort I added to make this software even more useful from an academic standpoint.
