### 2019-09-23 #1752 @nnadeau
@KingZee please see my comments below:

- In the `Solution space` section of the paper, I am assuming that you are describing your `QuickSolver` approach, but this is not explicit. Only the solution space is described, not the heuristic algorithm approach beyond for the exclusion of zero-time operations.
- There is no comparison of this heuristic approach to other state-of-the-art algorithms. A quick Google Scholar search for `makespan heuristic` reveals many many algorithms. How is your algorithm different? Why is it better than the existing literature?
- I respectfully disagree that the repository is well-documented; there is no API reference for developers to review (e.g., `doxygen`, `Read the Docs`).
- Java provides two key concepts for extensible and collaborative design: [`Interfaces`](https://docs.oracle.com/javase/tutorial/java/concepts/interface.html) and [`Access Control`](https://docs.oracle.com/javase/tutorial/java/javaOO/accesscontrol.html).
  - There isn't a single usage of the Java interface model.
  - Without proper API documentation, the public/private/protected members are not clear.
- Through the paper, the `README`, or the statement of need, I am not convinced why a developer should use and extend this project as opposed to starting from scratch. The is especially true when reviewing the code health/quality, for example:
  - The usage of `Thread.sleep();`.
  - The usage of `System.out.println()` for logging instead of a proper logging framework.
  - Direct calls to the garbage collector: `System.gc()`
  - The inclusion of `javafx` in nearly every class is a heavy dependency
  - The `Solver` class is a loose collection of some helper functions and a `JobData` container.
- Simply being able to solve shop optimization is not a valid minimum requirement for publishing a paper and repository.
