### 2019-09-25 #1752 @nthomasCUBE
Hi guys,

so far i still struggle with compiling and running,
managed to use jre11 but now have to manage to make it run,..
then I will provide the comments probably in the next week or so hopefully
