### 2019-09-25 #1752 @gkthiruvathukal
Hi, I plan to comment after both reviewers chime in.

Just a reminder that the purpose of review is to identify issues leading to improvements by the authors. So please focus on concrete issues and use the issue tracker. Many reviewers open up specific issues using the issue tracker. Once these are addressed (possibly with back and forth with author) we can move forward.

The goal is to ensure quality of the accepted submission. Hope that helps.
