### 2019-09-30 9c85e8c
Added link for JDK 11.
