+++ README.md
@@ -9,10 +9,9 @@ Problem and solution space encoding schemes are optimized across all three types
 
 It is extremely easy to implement any "Solver" for any of these types by inheriting from the base abstract Solver class and overriding corresponding methods.
 
-
 ## Dependencies :
 
-This project is built with JDK11 & Maven. It uses JavaFX 11 with some [JFoenix 9](https://github.com/jfoenixadmin/JFoenix) components for UI. JMH toolkit is included for micro-benchmarks.
+This project is built with JDK11 & Maven. It uses JavaFX 11 with some [JFoenix 9](https://github.com/jfoenixadmin/JFoenix) components for UI. JMH toolkit is included for micro-benchmarks. JDK 11 can be downloaded from the OpenJDK archive here : https://jdk.java.net/archive/.
 
 ## Installation :
 
