### 2019-09-30 #1752 @KingZee
Hi @nthomasCUBE, thanks a lot for the comments. I have made a new joss branch, so as to get everything working there before merging with master.

> SOFTWARE aspect
>
> -> authors
>
> i agree that role of coauthors should be stated somewhere. Maybe they can be added into the github project.

It's true that they are not contributors to the development of the project, but both of them have proof-read and partially co-authored the paper. I could ask them to make github accounts and make symbolic contributions if absolutely necessary. Maybe @gkthiruvathukal can chime in.

> -> installation
> Tool can be run easily with the description described on github. However, it does not work under the latest Java version. Maybe on can describe how the tool can be build and started in Eclipse where '--module-path C:/Users/thoma/Desktop/javafx-sdk-11.0.2/lib --add-modules javafx.controls,javafx.fxml' has to be put in 'Run Configuration'->'Arguments'->'VM arguments'
> and maybe on can link directly to the jdk-11.0.4 download site.

Thank you, I have made n edit to the readme added a link to JDK 11. (all of the changes are going to be in the JOSS branch for now). Also since each compiler has their own VM arguments prompt, I feel that including explanation on the exact steps for how to add the VM arguments isn't too relevant, as the VM arguments themselves should be enough.

> -> software
>
> gui looks nice and easy to use, probably would be nice to have a possibility to upload a matrix instead of manually adding values.

This is a very good feature request. I will try to implement it as soon as I have the time.

> -> comparison to other tools missing
>
> would be nice to list which other tools do similar analysis and whether the
> calculation time is better/worse etc.

As I mentioned before, I have found it very difficult to find software with a goal of scheduling jobs, when I looked for software that solves all job-shop, open-shop, and flow-shop problems I couldn't find any at all. There are some software that simply work as "GUI wrapper"s, and do not provide any solving framework. I have still found some software that generate Gantt charts for the job-shop variant. The motive of this software was in fact the lack of such a framework. There are algorithm implementations and libraries such as (https://github.com/fengelhardt/LiSA), with a large number of algorithms written across years of research, but there isn't that framework to help speed up setup time, as from the literature that I've read it simply seems each researcher re-writes their own implementation of shop scheduling from scratch, and only then proceed to code the solving part.

I reckon I still have to do my due diligence and make an exhaustive in-depth search of all software used here, and give a final answer backed with references.

> PAPER aspect (not done yet)
>
> p1. NP-hard, maybe to add a ref that these problems are NP-hard
> p2. would like to see on the github page an explaination how own algorithms can be included into the tool

I will add a section for contribution and explain extension of the Solver class.

I apologize if this reply is a bit late, comment edits don't trigger a notification so I just noticed your progress as I checked on this. It would be convenient if you raised an issue or made a new comment here for future review comments you might have. Thank you for your effort.
