### 2019-09-30 #1752 @gkthiruvathukal
All, ok, now that there is reviewer feedback (thanks @nnadeau and @nthomasCUBE) and a response from @KingZee, we are heading in the right direction. Please keep in mind that JOSS operates differently from your typical academic journal. As an editor, I rarely put anything up for review that lacks some novelty or technical merit, and I see potential in this work. Ultimately,  the formal peer review process is designed to improve the quality of the software submitted. I'm therefore going to focus on the quality issues. Being completely novel is not a prerequisite for JOSS submissions but clear motivation for why this is research software and having an audience (statement of need) is. Therefore, we will need to keep these in mind as we move forward.

I'm going to attempt to synthesize but here is a checklist of things I think need to be done before I can recommend acceptance, which is a union of the reviews + my own additions.

- [ ] **Add any and all additional contributors.** I do not hold the view that one must make formal commits to a repository to be a contributor. Nevertheless, I do think all contributors should seriously consider getting GitHub IDs and being collaborators on the repository. Among other things, this can help with long-term sustainability.

- [ ] **Java build issues need to be addressed.** I'd really like to see a continuous integration system used here (e.g. Travis, or perhaps the new GitHub Actions) to show a successful build and running the unit tests, especially on future commits. Java is a rapidly moving target now, and one hopes this software will continue to work in newer versions of Java.

- [ ] **Documentation needs improvements per earlier feedback.** I've never been a huge fan of javadoc myself, but perhaps there is a way to organize it so the key interfaces are discussed in detail with many examples. Another possibility is just to build better docs within the repo itself, possibly with proper design diagrams or an overview of the code organization. The points about interfaces (an absolute must in professional Java programming), using proper logging frameworks, and Thread (concurrency) management need to be addressed.

- [ ] **Java code quality seems to be a valid concern.** I think to meet y'all in the middle, the best way to address this is to run one of the "code smells" tools. This journal article has a nice overview, https://jserd.springeropen.com/articles/10.1186/s40411-017-0041-1. Might be a bit out of date now, but at least you can start there to get an idea of the leading tools. Upon doing this analysis, you will probably want to do a bit of code refactoring.

- [ ] **Related work.** This is a must. There are other tools out there, I'm sure, but most probably don't have a GUI and--in some cases--are libraries. Please make sure to identify related work.

- [ ] **Statement of need.** Please make sure this is well-motivated. Ultimately, the hope is that the work gets cited, especially if it supports other research projects (your or others).

I don't think any of the above is impossible to address. While it may take some time, we can work with your schedule, @KingZee. As long as you are willing to see the submission through the process, we can eventually get to a proper accepting state.
