### 2019-09-30 #1752 @KingZee
@gkthiruvathukal Thank you very much for your bullet points, I absolutely agree with them and feel that they all are very pertinent. I will do my best over the course of the next few weeks to check off everything.

I have a single concern about the second half of the third bullet point. I agree with both of you and @nnadeau that javadoc might not have been the best choice,  and will do my best to re-write all of it in a repo-friendly manner.

But :
> The points about interfaces (an absolute must in professional Java programming), using proper logging frameworks, and Thread (concurrency) management need to be addressed.

These are points that might need to be clarified.

1. There is a total of 4 System.out.println lines in all of my classes. Three of them are simple flags that print "Event running in background", "Event failed", and "Event canceled", which is already superfluous as I am already handling every class of throwable errors in the failed event for example. I could simply delete the lines to absolutely no effect on the code, which really just makes me feel they were low-hanging nitpicks to judge the quality of my code. I will proceed to delete them, I just didn't expect them to be that much of a topic here. I apologize for thinking use of System.out is "okay", because apparently it is not.

2. Thread management is a point that might need to be discussed on a separate issue, I invite @nnadeau to open one in the repo. The use of Thread.sleep() was justified in my opinion as the heavy nature of the computation would take up so much resources it hinders any use of the machine otherwise. But again, this is up for discussion.

3. I still am heavily not convinced about the use of interfaces, especially because the change to them would almost be trivial : As I am currently using an abstract Solver base class that handles all the threading code, all I would need would to take out the corresponding methods into an interface, and make this Solver class implement it. But here is what I feel like I would be losing :

* While an interface defines the contract future developers will need to implement, an abstract class inheritance still creates a much stronger contract, I can demonstrate that through many examples :

- Using an interface opens up room for new parent classes, of which then more developers can branch out from. This is extremely harder to maintain than a single abstract Solver class that all solvers inherit from. This is a problem that is easily visible once this software has tens or dozens of algorithms, which each may have multiple variants.

- A base Solver class provides a ready-to-go "plug and play" kind of structure for other algorithms, to benefit from the GUI interface, from the background threading code, and from repetitive utility methods. It would take seconds to add more algorithms to the GUI assuming they inherit from the existing class, and since abstract classes can have bodies, any computation method that could be repetitive across multiple solvers (like the existing cartesian multiplication, schedule permutation and parsing, etc...) can be added to the abstract class.

- All future solvers will improve from common method improvements. If any of the methods I mentioned above are improved, every solver will have a performance boost. It is also true that this is a double-edged sword, as changes in these methods might also cause breakage. But that is why we have test suites.

- I already coded the abstract class with the utmost freedom in mind. The solver method not only allows the algorithm to return a single optimal solution, but a list of solutions. This is extremely useful for many heuristics, like Ant Colony Optimisation, or genetic/evolutionary algorithms, that work with populations. I can already see the implementation of those algorithms and more working seamlessly with the code I have now.

Trust me I have had this discussion with myself, other developers, and have convinced myself of the use of abstract classes in this specific scenario. I understand that interfaces are common ways to define such contracts, but please understand that this was a logical choice with future implementations in mind.

That would be all, I would like to thank @gkthiruvathukal again for the great synthesis, I'll get to working on it as soon as possible. I would also like to thank both reviewers for their work in judging my submission.
