### 2020-01-20 #1752 @gkthiruvathukal
@danielskatz @openjournals/joss-eics This submission has been paused since end of October. Should we go ahead and close the issue, for now?
