### 2020-02-08 ddcc9e8
Refactored & fixed PMD warnings.
