### 2020-02-08 KingZee/shop-solver#3 @KingZee
https://github.com/KingZee/shop-solver/commit/ddcc9e8eca05903fa9ef9755f04c4ccce6208186

Fixed all warnings.

Only remaining "UnusedPrivateMethod" and "UnusedFormalParameter" are false positives as they are event methods annotated by the `@FXML` directive
