### 2020-02-11 #1752 @KingZee
@danielskatz I have currently began working on the mentioned issues. https://github.com/openjournals/joss-reviews/issues/1752#issuecomment-536621094 This comment summarized what needs work, I closed an issue related to the 3rd point just a few days ago, https://github.com/KingZee/shop-solver/issues/3 . I'm currently figuring out better documentation, I used doxygen as recommended but I did not see any benefit over javadoc, so I am figuring out better alternatives. I will probably make a new commit on the issue soon.
