+++ .github/workflows/maven.yml
@@ -5,7 +5,7 @@ on: [push]
 jobs:
   build:
 
-    runs-on: ubuntu-latest
+    runs-on: windows-latest
 
     steps:
     - uses: actions/checkout@v2
@@ -15,4 +15,4 @@ jobs:
         java-version: 11
         java-package: jdk+fx
     - name: Build with Maven
-      run: xvfb-run mvn -B package --file pom.xml
+      run: mvn -B package --file pom.xml
