### 2020-02-15 e43a20b
Added Travis CI config.
