+++ .github/workflows/maven.yml
@@ -1,0 +1,18 @@
+name: Java CI
+
+on: [push]
+
+jobs:
+  build:
+
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v2
+      - name: Setup Java JDK
+        uses: actions/setup-java@v1.3.0
+        with:
+          java-version: 11
+          java-package: jdk+fx
+      - name: Build with Maven
+        run: mvn -B package --file pom.xml -Dtestfx.robot=glass -Dglass.platform=Monocle -Dmonocle.platform=Headless
