+++ .github/workflows/maven.yml
@@ -15,4 +15,6 @@ jobs:
           java-version: 11
           java-package: jdk+fx
       - name: Build with Maven
-        run: mvn -B package --file pom.xml -Dtestfx.robot=glass -Dglass.platform=Monocle -Dmonocle.platform=Headless
+        run: mvn clean install
+      - name: Run test suite
+        run: mvn test -Dtestfx.robot=glass -Dglass.platform=Monocle -Dmonocle.platform=Headless
