+++ .github/workflows/maven.yml
@@ -14,7 +14,7 @@ jobs:
         with:
           java-version: 11
           java-package: jdk+fx
+      - name : Install display library
+        run: sudo apt-get install libpangoft2-1.0
       - name: Build with Maven
-        run: mvn clean compile
-      - name: Run test suite
-        run: mvn test -Dtestfx.robot=glass -Dglass.platform=Monocle -Dmonocle.platform=Headless
+        run: mvn -B package --file pom.xml -Dtestfx.robot=glass -Dglass.platform=Monocle -Dmonocle.platform=Headless -Dprism.order=sw
