+++ .github/workflows/maven.yml
@@ -9,6 +9,8 @@ jobs:
 
     steps:
       - uses: actions/checkout@v2
+        with:
+          ref: github-action
       - name: Setup Java JDK
         uses: actions/setup-java@v1.3.0
         with:
