+++ .github/workflows/maven.yml
@@ -18,5 +18,7 @@ jobs:
           java-package: jdk
       - name : Install display library
         run: sudo apt-get install libpangoft2-1.0
+      - name: Checkout
+        run: git checkout joss
       - name: Build with Maven
         run: mvn -B package --file pom.xml -Dtestfx.robot=glass -Dglass.platform=Monocle -Dmonocle.platform=Headless -Dprism.order=sw
