+++ .github/workflows/maven.yml
@@ -5,7 +5,7 @@ on: [push]
 jobs:
   build:
 
-    runs-on: ubuntu-latest
+    runs-on: windows-latest
 
     steps:
       - uses: actions/checkout@v2
@@ -15,7 +15,5 @@ jobs:
         uses: actions/setup-java@v1
         with:
           java-version: 11.0.6
-      - name : Install display library
-        run: sudo apt-get install libpangoft2-1.0
       - name: Build with Maven
-        run: mvn -B package --file pom.xml -Dtestfx.robot=glass -Dglass.platform=Monocle -Dmonocle.platform=Headless -Dprism.order=sw -Dprism.text=t2k
+        run: mvn -B package --file pom.xml -Dtestfx.robot=glass -Dglass.platform=Monocle
