+++ .github/workflows/maven.yml
@@ -13,11 +13,11 @@ jobs:
           ref: github-action
       - name: Install JDK
         run: |
-        sudo apt-get install openjdk-11-jdk
-        sudo apt-get install maven
+          sudo apt-get install openjdk-11-jdk
+          sudo apt-get install maven
       - name : Install display library
         run: sudo apt-get install libpangoft2-1.0
       - name: Build with Maven
         run: |
-        mvn -version
-        mvn -B package --file pom.xml -Dtestfx.robot=glass -Dglass.platform=Monocle -Dmonocle.platform=Headless -Dprism.order=sw -Dprism.text=t2k
+          mvn -version
+          mvn -B package --file pom.xml -Dtestfx.robot=glass -Dglass.platform=Monocle -Dmonocle.platform=Headless -Dprism.order=sw -Dprism.text=t2k
