+++ .github/workflows/maven.yml
@@ -11,10 +11,10 @@ jobs:
       - uses: actions/checkout@v2
         with:
           ref: github-action
-      - name: Setup Java JDK
-        uses: actions/setup-java@v1
-        with:
-          java-version: 11.0.6
+      - name: Install JDK
+        run: |
+          sudo apt-get install oracle-java11-set-default-local
+          sudo apt-get install maven
       - name : Install display library
         run: sudo apt-get install libpangoft2-1.0
       - name: Build with Maven
