+++ .github/workflows/maven.yml
@@ -12,13 +12,12 @@ jobs:
         with:
           ref: github-action
       - name: Install JDK
-        run: |
-          sudo apt-get install openjdk-11-jdk
-          sudo apt-get install maven
-          export JAVA_HOME=/usr/lib/jvm/java-11-openjdk-amd64
+        run: sudo apt-get install openjdk-11-jdk maven && export JAVA_HOME=/usr/lib/jvm/java-11-openjdk-amd64
       - name : Install display library
         run: sudo apt-get install libpangoft2-1.0
       - name: Build with Maven
         run: |
+          echo $JAVA_HOME
+          export JAVA_HOME=/usr/lib/jvm/java-11-openjdk-amd64
           mvn -version
           mvn -B package --file pom.xml -Dtestfx.robot=glass -Dglass.platform=Monocle -Dmonocle.platform=Headless -Dprism.order=sw -Dprism.text=t2k
