### 2020-02-17 1a365e3
Updated Github action.
