### 2020-02-17 0b92138
Merge github action into joss branch (#6)
* Added Github action.
