### 2020-02-17 37a04d4
Updated Readme with Build status badge.
